restapi/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
numpy==1.22.4
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9